Utilties/galaxy-vault/pyproject.toml
[tool.poetry]
name = "galaxy-vault"
version = "0.1.5"
description = "Internal Package to fetch necessary crednetials and secrest to access various resources for Galaxy apps. Currently utilizes Azure Key Vault or Local Config to fetch secrets."
authors = ["mnguyenE1001 <matthew.nguyen@external.engie.com>"]
readme = "README.md"
packages = [{include = "galaxy_vault"}]

[tool.poetry.dependencies]
python = "^3.10"
azure-keyvault-secrets = "^4.7.0"
azure-identity = "^1.15.0"
pyyaml = "^6.0.1"
pydash = "^7.0.7"
azure-keyvault-certificates = "^4.7.0"
cryptography = "^42.0.2"
pip-system-certs = "^4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
